@@ -1,5 +1,5 @@
 import asyncio
-import asynctest
+import async_solipsism
 from unittest import mock
 
 import katsdptelstate.aio
@@ -8,6 +8,8 @@ import kattelmod
 from kattelmod.component import Component, TelstateUpdatingComponent, KATCPComponent, MultiMethod
 from kattelmod.clock import get_clock
 from kattelmod.test.test_clock import WarpEventLoopTestCase
+import re
+import pytest
 
 
 class DummyComponent(Component):
@@ -30,19 +32,17 @@ class DummyTelstateUpdatingComponent(TelstateUpdatingComponent):
 
 
 class TestComponent(WarpEventLoopTestCase):
-    def setUp(self):
+    def setup_method(self):
         self.comp = DummyComponent(88.0)
 
     def test_type(self):
-        self.assertEqual(self.comp._type(), 'kattelmod.test.test_component.DummyComponent')
+        assert self.comp._type() == 'kattelmod.test.test_component.DummyComponent'
 
     def test_repr(self):
-        self.assertRegex(
-            repr(self.comp),
-            r"<kattelmod\.test\.test_component\.DummyComponent 'dummy' at .*>")
+        assert re.search(r"<kattelmod\.test\.test_component\.DummyComponent 'dummy' at .*>", repr(self.comp))
 
     def test_sensors(self):
-        self.assertEqual(self.comp._sensors, ['pressure', 'speed', 'temperature'])
+        assert self.comp._sensors == ['pressure', 'speed', 'temperature']
 
     async def test_dummy_methods(self):
         # Just tests that it exists and runs without crashing
@@ -57,13 +57,15 @@ class TestComponent(WarpEventLoopTestCase):
                         return_value=kattelmod.test.test_component):
             self.comp.temperature = 100.0
             fake = self.comp._fake()
-            self.assertEqual(fake.temperature, 100.0)
-            self.assertEqual(fake.pressure, 1000.0)
-            self.assertEqual(fake.speed, 88.0)
+            assert fake.temperature == 100.0
+            assert fake.pressure == 1000.0
+            assert fake.speed == 88.0
 
 
 class TestTelstateUpdatingComponent(WarpEventLoopTestCase):
-    async def setUp(self):
+    # XXX Make it autouse until Pytest supports async setup_methods
+    @pytest.fixture(autouse=True)
+    async def setup_method(self):
         self.telstate = katsdptelstate.aio.TelescopeState()
         self.comp = DummyTelstateUpdatingComponent(88.0)
         self.comp._name = 'dummy'
@@ -76,17 +78,17 @@ class TestTelstateUpdatingComponent(WarpEventLoopTestCase):
     async def test_setattr(self):
         """Basic test that setting an attribute updates telstate"""
         await self.comp._flush()
-        self.assertEqual(await self.telstate.get('dummy_speed'), 88.0)
-        self.assertEqual(await self.telstate.key_type('dummy_speed'),
-                         katsdptelstate.KeyType.IMMUTABLE)
+        assert await self.telstate.get('dummy_speed') == 88.0
+        assert await self.telstate.key_type('dummy_speed') == \
+                         katsdptelstate.KeyType.IMMUTABLE
         # Initial value is pushed into the past
-        self.assertEqual(await self.telstate.get_range('dummy_temperature', st=0),
-                         [(451.0, self.START_TIME - 300.0)])
+        assert await self.telstate.get_range('dummy_temperature', st=0) == \
+                         [(451.0, self.START_TIME - 300.0)]
         get_clock().advance(5)
         self.comp.temperature = 100.0
         await self.comp._flush()
-        self.assertEqual(await self.telstate.get_range('dummy_temperature', st=0),
-                         [(451.0, self.START_TIME - 300.0), (100.0, self.START_TIME + 5.0)])
+        assert await self.telstate.get_range('dummy_temperature', st=0) == \
+                         [(451.0, self.START_TIME - 300.0), (100.0, self.START_TIME + 5.0)]
 
     async def test_updates(self):
         """Test interaction with _update and rate limiting.
@@ -101,55 +103,65 @@ class TestTelstateUpdatingComponent(WarpEventLoopTestCase):
             self.comp.temperature += 1.0
             self.comp.pos_foo += 1.0
         await self.comp._flush()
-        self.assertEqual(
-            await self.telstate.get_range('dummy_temperature', st=0),
+        assert await self.telstate.get_range('dummy_temperature', st=0) == \
             [(451.0, self.START_TIME - 300.0),
              (452.0, self.START_TIME + 0.25),
              (453.0, self.START_TIME + 0.5),
              (454.0, self.START_TIME + 0.75),
              (455.0, self.START_TIME + 1.0),
              (456.0, self.START_TIME + 1.25),
-             (457.0, self.START_TIME + 1.5)])
-        self.assertEqual(
-            await self.telstate.get_range('dummy_pos_foo', st=0),
+             (457.0, self.START_TIME + 1.5)]
+        assert await self.telstate.get_range('dummy_pos_foo', st=0) == \
             [(1000.0, self.START_TIME - 300.0),
              (1002.0, self.START_TIME + 0.5),
              (1004.0, self.START_TIME + 1.0),
-             (1006.0, self.START_TIME + 1.5)])
+             (1006.0, self.START_TIME + 1.5)]
 
     def test_updatable(self):
-        self.assertTrue(self.comp._updatable)
+        assert self.comp._updatable
 
     async def test_start_twice(self):
         # Mostly just to get test coverage
         await self.comp._start()
 
 
-class TestKATCPComponent(asynctest.ClockedTestCase):
-    async def setUp(self) -> None:
+def test_bad_katcp_endpoint():
+    with pytest.raises(ValueError):
+        KATCPComponent('invalid.com')
+    with pytest.raises(ValueError):
+        KATCPComponent('')
+
+
+class TestKATCPComponent:
+    @classmethod
+    @pytest.fixture
+    def event_loop(cls):
+        loop = async_solipsism.EventLoop()
+        yield loop
+        loop.close()
+
+    @pytest.fixture(autouse=True)
+    async def setup_method(self) -> None:
+        loop = asyncio.get_running_loop()
         self.server = await asyncio.start_server(self._client_cb, '127.0.0.1', 0)
         self.endpoint = self.server.sockets[0].getsockname()[:2]
-        self.addCleanup(self.server.wait_closed)
-        self.addCleanup(self.server.close)
-        self.reader = self.loop.create_future()    # type: asyncio.Future[asyncio.StreamReader]
-        self.writer = self.loop.create_future()    # type: asyncio.Future[asyncio.StreamWriter]
+        self.reader = loop.create_future()    # type: asyncio.Future[asyncio.StreamReader]
+        self.writer = loop.create_future()    # type: asyncio.Future[asyncio.StreamWriter]
+        yield
+        writer = await self.writer
+        writer.close()
+        self.server.close()
+        await self.server.wait_closed()
 
     def _client_cb(self, reader: asyncio.StreamReader, writer: asyncio.StreamWriter) -> None:
         self.reader.set_result(reader)
         self.writer.set_result(writer)
-        self.addCleanup(writer.close)
-
-    def test_bad_endpoint(self):
-        with self.assertRaises(ValueError):
-            KATCPComponent('invalid.com')
-        with self.assertRaises(ValueError):
-            KATCPComponent('')
 
     async def test_connect_timeout(self):
         comp = KATCPComponent('{}:{}'.format(*self.endpoint))
         task = asyncio.ensure_future(comp._start())
-        await self.advance(100)
-        with self.assertRaises(asyncio.TimeoutError):
+        await asyncio.sleep(100)
+        with pytest.raises(asyncio.TimeoutError):
             await task
 
     async def _interact(self):
@@ -162,12 +174,13 @@ class TestKATCPComponent(asynctest.ClockedTestCase):
         await writer.drain()
 
     async def test_good(self):
-        task = self.loop.create_task(self._interact())
+        loop = asyncio.get_running_loop()
+        task = loop.create_task(self._interact())
         comp = KATCPComponent('{}:{}'.format(*self.endpoint))
         await comp._start()
         await comp._start()      # Check that it's idempotent
         response = await comp._client.request('ping', 'hello')
-        self.assertEqual(response, ([b'hello'], []))
+        assert response == ([b'hello'], [])
         await comp._stop()
         await comp._stop()       # Check that it's idempotent
         await task
@@ -190,8 +203,8 @@ class AsyncMethod:
         self.called_with = (args, kwargs)
 
 
-class TestMultiMethod(asynctest.TestCase):
-    def setUp(self):
+class TestMultiMethod:
+    def setup_method(self):
         self.sync1 = SyncMethod()
         self.sync2 = SyncMethod()
         self.async1 = AsyncMethod()
@@ -199,16 +212,16 @@ class TestMultiMethod(asynctest.TestCase):
 
     def test_sync(self):
         mm = MultiMethod([self.sync1, self.sync2], 'my_method', 'help')
-        self.assertIsNone(self.sync1.called_with)
-        self.assertIsNone(self.sync2.called_with)
+        assert self.sync1.called_with is None
+        assert self.sync2.called_with is None
         mm('a', 1, kw=2)
-        self.assertEqual(self.sync1.called_with, (('a', 1), {'kw': 2}))
-        self.assertEqual(self.sync2.called_with, (('a', 1), {'kw': 2}))
+        assert self.sync1.called_with == (('a', 1), {'kw': 2})
+        assert self.sync2.called_with == (('a', 1), {'kw': 2})
 
     async def test_async(self):
         mm = MultiMethod([self.async1, self.async2], 'my_method', 'help')
-        self.assertIsNone(self.async1.called_with)
-        self.assertIsNone(self.async2.called_with)
+        assert self.async1.called_with is None
+        assert self.async2.called_with is None
         await mm('a', 1, kw=2)
-        self.assertEqual(self.async1.called_with, (('a', 1), {'kw': 2}))
-        self.assertEqual(self.async2.called_with, (('a', 1), {'kw': 2}))
+        assert self.async1.called_with == (('a', 1), {'kw': 2})
+        assert self.async2.called_with == (('a', 1), {'kw': 2})
